+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/subjects/movie/MovieScreen.kt
@@ -76,7 +76,7 @@ fun MovieScreen(
     showCreateDouListDialog: Boolean,
     onBackClick: () -> Unit,
     onLoginClick: () -> Unit,
-    onUpdateStatus: (newStatus: SubjectInterestStatus) -> Unit,
+    onUpdateStatus: (newStatus: SubjectInterestStatus, rating: Int?) -> Unit,
     onImageClick: (url: String) -> Unit,
     onUserClick: (userId: String) -> Unit,
     onRecommendSubjectClick: (subject: RecommendSubject) -> Unit,