+++ app/src/main/java/com/github/bumblebee202111/doubean/model/subjects/SubjectInterest.kt
@@ -2,6 +2,7 @@ package com.github.bumblebee202111.doubean.model.subjects
 
 data class SubjectInterest(
     val comment: String?,
+    val rating: Rating?,
     val status: SubjectInterestStatus,
 )