+++ app/src/main/java/com/github/bumblebee202111/doubean/model/subjects/SubjectWithInterest.kt
@@ -3,8 +3,9 @@ package com.github.bumblebee202111.doubean.model.subjects
 data class SubjectWithInterest<T : Subject>(
     val subject: T,
     override val interest: SubjectInterest = SubjectInterest(
-        null,
-        SubjectInterestStatus.MARK_STATUS_UNMARK
+        comment = null,
+        rating = null,
+        status = SubjectInterestStatus.MARK_STATUS_UNMARK
     ),
 ) : MarkableSubject {
     override val id: String