+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Change Log
 
 ## Version 0.11.0-beta1
 
+## Version 0.11.0-beta1
+
+- **New:** You can now rate subjects (Movies, Books, TVs) when marking them as "Doing"/"Done". Your
+  personal rating is also displayed in the subject header. (
+  Addresses [#31](https://github.com/Bumblebee202111/doubean-public/issues/31))
+
 ## Version 0.10.3
 
 - **Fixed:** Image saving now works correctly on Android 10 devices. (